Ajustando o fluxo para salvar o detalhe do item

+++ lib/features/shopping/bloc/create_detail_item_shoppinglist_event.dart
@@ -19,7 +19,7 @@ class FetchDetailItemShoppingListRequested extends CreateDetailItemShoppinglistE
 class DetailItemUpdateShoppingListRequested extends CreateDetailItemShoppinglistEvent {
   final String itemId;
   final String? name;
-  final int? quantity;
+  final double? quantity;
   final double? price;
   final bool? checked;
